fixed BitVector_
git-svn-id: http://svn.seqan.de/seqan/trunk@10213 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/include/seqan/basic/tuple_bit_compressed.h
@@ -72,7 +72,7 @@ struct BitVector_
 
 template <> struct BitVector_<8> { typedef unsigned char Type; };
 template <> struct BitVector_<16> { typedef unsigned short Type; };
-template <> struct BitVector_<32> { typedef unsigned long Type; };
+template <> struct BitVector_<32> { typedef unsigned int Type; };
 template <> struct BitVector_<64> { typedef __uint64 Type; };
 template <> struct BitVector_<255> { typedef __uint64 Type; };
